MetricsRecord to carry long value, metric's timeseries table to store long values

+++ watchdog/src/main/java/co/cask/cdap/metrics/data/AggregatesTable.java
@@ -107,7 +107,7 @@ public final class AggregatesTable {
 
         // For each tag, increments corresponding values
         for (TagMetric tag : record.getTags()) {
-          increments.put(Bytes.toBytes(tag.getTag()), (long) tag.getValue());
+          increments.put(Bytes.toBytes(tag.getTag()), tag.getValue());
         }
         aggregatesTable.increment(rowKey, increments);
       }
